requirements.txt
pycodestyle==2.7.0
discord.py==1.7.3
python-dotenv==0.19.0
requests==2.26.0
types-requests==2.25.9
typing-extensions==3.10.0.2
pysqlite3==0.4.6 